requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.2.0
alabaster==0.7.13
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.12.1
backcall==0.2.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.4
chex==0.1.7
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==1.6.0
cmake==3.27.0
comm==0.2.1
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
deepdiff==6.3.1
dm-tree==0.1.8
docutils==0.18.1
et-xmlfile==1.1.0
etils==1.3.0
execnet==2.0.2
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.12.2
Fiona==1.9.4.post1
flatbuffers==23.5.26
flax==0.7.2
fonttools==4.25.0
frozenlist==1.3.3
gast==0.4.0
geopandas==0.13.2
gmpy2==2.1.2
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
Grid2Op==1.9.8
grpcio==1.60.0
gymnasium==0.28.1
h5py==3.6.0
idna==3.3
igraph==0.10.6
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.12.3
ipywidgets==8.1.1
jax==0.4.13
jax-jumpy==1.0.0
jaxlib==0.4.13
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==16.0.6
LightSim2Grid==0.7.5
lit==16.0.6
llvmlite==0.39.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
matpowercaseframes==1.0.6
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.2
munkres==1.1.4
nest-asyncio==1.5.8
networkx==3.1
numba==0.56.4
numexpr==2.8.4
numpy==1.22.3
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
openpyxl==3.1.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.3
ordered-set==4.1.0
ortools==9.7.2996
packaging==21.3
pandapower==2.13.1
pandas==1.3.5
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
pip==23.2.1
plotly==5.16.1
pluggy==1.2.0
ply==3.11
pooch==1.4.0
prompt-toolkit==3.0.43
protobuf==4.23.3
psutil==5.9.0
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.16.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyproj==3.5.0
PyQt5-sip==12.11.0
PySocks==1.7.1
pytest==7.0.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-igraph==0.10.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rich==13.7.0
rsa==4.7.2
sb3-contrib==2.2.1
sbx-rl==0.9.1
scipy==1.7.3
setuptools==67.8.0
shapely==2.0.1
sip==6.6.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stable-baselines3==2.2.1
stack-data==0.6.3
sympy==1.9
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.21.0
tensorstore==0.1.45
termcolor==1.1.0
texttable==1.6.7
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.14.1
triton==2.0.0
typing_extensions==4.3.0
urllib3==1.26.8
wcwidth==0.2.13
Werkzeug==2.0.3
wheel==0.38.4
widgetsnbextension==4.0.9
wrapt==1.13.3
XlsxWriter==3.1.2
yarl==1.8.1
zipp==3.7.0
